+++ NEWS.md
@@ -1,4 +1,22 @@
-# eegUtils 0.6.0.9999
+# eegUtils 0.6.2
+
+### Function changes
+- added support for `EEGLAB` .set files saved in newer Matlab file formats.
+- changed first argument of `eeg_filter()` to `data` instead of `.data`
+- added some more informative user messages for importing data and adding electrode locations.
+- `erp_image()` now supports `eeg_tfr` objects.
+
+### Internal changes / bug fixes
+
+- When combining three or more continuous `eeg_data` objects, `eeg_combine()` would substantially undercorrect the timing of events in the third file - this is now fixed.
+- `groups` parameter for `topoplot()` now correctly passed for all types of object.
+- `stat_tests.R` file removed, will be reimplemented elsewhere
+- Long standing issues with import of channel locations from EEGLAB files hopefully fixed...
+- `rm_baseline()` for `eeg_evoked` no longer uses `data.table`
+- `as.data.frame.eeg_evoked()` handles grouped data better.
+- `import_set()` now handles all EEGLAB formats better.
+
+# eegUtils 0.6.1
 
 ### Function changes